+++ src/app/app-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { BudgetModule } from './budget/budget.module';
-import { HomeComponent } from './home/home.component';
+import { BudgetModule } from './modules/budget.module';
+import { HomeComponent } from './components/home/home.component';
 
 const routes: Routes = [
   { path: 'home', component: HomeComponent },
-  { path: 'pfinance/budget', loadChildren: () => import('./budget/budget.module').then(m => m.BudgetModule)  },
+  { path: 'pfinance/budget', loadChildren: () => import('./modules/budget.module').then(m => m.BudgetModule)  },
   { path: '', redirectTo: 'home', pathMatch: 'full' },
   { path: '**', redirectTo: 'home', pathMatch: 'full' }
 ];
