+++ src/app/app.module.ts
@@ -5,8 +5,8 @@ import { BrowserModule } from '@angular/platform-browser';
 import { HttpClientModule } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
-import { HomeComponent } from './home/home.component';
-import { SignUpModalComponent } from './modals/signUpModal/signUpModal.component';
+import { HomeComponent } from './components/home/home.component';
+import { SignUpModalComponent } from './components/modals/signUpModal/signUpModal.component';
 import { AppRoutingModule } from './app-routing.module';
