+++ src/app/components/modals/signUpModal/signUpModal.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { ModalService } from '../modal.service';
+import { ModalService } from '../../../services/modal/modal.service';
 
 describe('ModalService', () => {
   let service: ModalService;
